+++ rhine-terminal/rhine-terminal.cabal
@@ -2,7 +2,7 @@
 -- documentation, see http://haskell.org/cabal/users-guide/
 
 name:                rhine-terminal
-version:             0.7.1
+version:             0.8.0
 synopsis:            Terminal backend for Rhine
 description:
   This package provides an example of a `terminal` based program using rhine.
@@ -24,7 +24,7 @@ source-repository head
 source-repository this
   type:     git
   location: git@github.com:turion/rhine.git
-  tag:      v0.7.1
+  tag:      v0.8.0.0
 
 library
   exposed-modules:
@@ -32,7 +32,7 @@ library
   build-depends:       base         >= 4.11 && < 4.16
                      , exceptions   >= 0.10.4
                      , transformers >= 0.5
-                     , rhine        == 0.7.1
+                     , rhine        == 0.8.0.0
                      , dunai        >= 0.6
                      , terminal     >= 0.2.0.0
                      , time         >= 1.9.3
@@ -40,14 +40,12 @@ library
   default-language:    Haskell2010
   ghc-options:         -Wall
 
-executable rhine-terminal-repl
-  main-is:             Main.hs
+executable rhine-terminal-simple
+  main-is:             TerminalSimple.hs
   ghc-options:         -threaded
   build-depends:       base         >= 4.11 && < 4.16
-                     , rhine        == 0.7.1
+                     , rhine        == 0.8.0.0
                      , rhine-terminal
-                     , exceptions   >= 0.10.4
-                     , transformers >= 0.5
                      , terminal     >= 0.2.0.0
                      , text         >= 1.2.5.0
                      , time         >= 1.9.3
@@ -60,7 +58,7 @@ test-suite rhine-terminal-tests
   main-is:             tests/Main.hs
   ghc-options:         -threaded
   build-depends:       base         >= 4.11 && < 4.16
-                     , rhine        == 0.7.1
+                     , rhine        == 0.8.0.0
                      , rhine-terminal
                      , exceptions   >= 0.10.4
                      , transformers >= 0.5
